pyproject.toml
[tool.poetry]
name = "fyreplace-api"
version = "0.0.0"
description = "Fyreplace API"
authors = []
license = "AGPL-3.0-or-later"
readme = "README.md"
packages = []

[tool.poetry.dependencies]
black = "^23.1"
celery = {extras = ["redis"], version = "^5.2"}
daphne = "^4.0"
django = {extras = ["argon2"], version = "^4.1"}
django-anymail = {extras = ["mailgun"], version = "^9.0"}
django-celery-beat = "^2.4"
django-extensions = "^3.2"
django-health-check = "^3.17"
django-storages = {extras = ["boto3"], version = "^1.13"}
django-stubs = "^1.15"
djangorestframework = "^3.14"
dj-database-url = "^1.2"
drf-spectacular = "^0.25"
email-validator = "^1.3"
firebase-admin = "^6.1"
httpx = {extras = ["http2"], version = "^0.23"}
ipython = "^8.10"
isort = "^5.12"
mysqlclient = "^2.1"
pillow = "^9.4"
psycopg2 = "^2.9"
pyjwt = {extras = ["crypto"], version = "^2.6"}
pytest-xdist = {extras = ["psutil"], version = "^3.2"}
pytest-django = "^4.5"
python = "^3.11"
python-dotenv = "^0.21"
python-magic = "^0.4"
qrcode = "^7.4"
requests = "^2.28"
tqdm = "^4.64"
sentry-sdk = "^1.15.0"
whitenoise = {extras = ["brotli"], version = "^6.3"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
